# order-service/src/main/resources/openapi/order-service.yml
openapi: 3.0.3
info:
  description: Open API specification for Order service
  version: 0.0.1
  title: Order Api Service Specification
  contact:
    name: JAVA developer
    email: palivtom@fel.cvut.cz
tags:
  - name: Order
  - name: Cart
paths:
  /orders:
    get:
      tags:
        - Order
      summary: Get orders
      operationId: getOrders
      responses:
        '200':
          description: Returns list of feedbacks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Order
      summary: Create new order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '201':
          description: Returns created order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /orders/{orderId}:
    get:
      tags:
        - Order
      summary: Get order by id
      operationId: getOrderById
      parameters:
        - in: path
          required: true
          name: orderId
          description: Id of searched order
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Returns order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Order
      summary: Update order by id
      operationId: updateOrder
      parameters:
        - in: path
          required: true
          name: orderId
          description: Id of updating order
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Returns updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Order
      summary: Cancel order by id
      operationId: cancelOrder
      parameters:
        - in: path
          required: true
          name: orderId
          description: Id of searched order
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Returns canceled order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /cart:
    get:
      tags:
        - Cart
      summary: Get cart
      operationId: getCart
      responses:
        '200':
          description: Returns cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /cart/add:
    post:
      tags:
        - Cart
      summary: Add product to cart
      operationId: addProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItem'
      responses:
        '200':
          description: Returns updated cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /cart/remove:
    post:
      tags:
        - Cart
      summary: Remove product from cart
      operationId: removeProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItem'
      responses:
        '200':
          description: Returns updated cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  responses:
    BadRequest:
      description: Indicates a nonspecific client error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Exception'

    Unauthorized:
      description: Unauthorized request - sent when the client either provided invalid credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Exception'

    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Exception'

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Exception'

  schemas:
    Exception:
      type: object
      properties:
        code:
          type: integer
          example: 400
        title:
          type: string
          example: Wrong input parameter
        message:
          type: string
          example: Wrong parameter for endpoint '/example/endpoint', wrong endpointId value.

    Order:
      type: object
      properties:
        id:
          readOnly: true
          type: integer
          format: int64
        userId:
          readOnly: true
          type: string
        shippingAddress:
          $ref: '#/components/schemas/ShippingAddress'
        createdAt:
          readOnly: true
          type: integer
          format: int64
        updatedAt:
          readOnly: true
          type: integer
          format: int64
        deletedAt:
          readOnly: true
          type: integer
          format: int64
          description: If order is canceled, this field is filled with date and time of canceling

    ShippingAddress:
      type: object
      required:
        - country
        - city
        - street
        - zipCode
      properties:
        country:
          type: string
        city:
          type: string
        street:
          type: string
        zipCode:
          type: string

    CartItem:
      type: object
      required:
        - productNo
        - quantity
      properties:
        productNo:
          type: string
        quantity:
          type: integer

    OrderView:
      type: object
      readOnly: true
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: string
        cart:
          $ref: '#/components/schemas/CartView'
        shippingAddress:
          $ref: '#/components/schemas/ShippingAddress'
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
        deletedAt:
          type: integer
          format: int64
          description: If order is canceled, this field is filled with date and time of canceling

    CartView:
      type: object
      readOnly: true
      properties:
        userId:
          type: string
        priceSum:
          type: integer
        itemCount:
          type: integer